move interfaces to deps.go
